rename app-bar to app-shell
