+++ src/components/index.js
@@ -1,4 +1,4 @@
-export { default as AppBar } from './app-bar/app-bar';
+export { default as AppShell } from './app-shell/app-shell';
 export { default as FooterBar } from './footer-bar/footer-bar';
 export { default as HeroElement } from './hero-element/hero-element';
 export { default as RatingBar } from './rating-bar/rating-bar';
